cbb-upsets/pyproject.toml
[project]
name = "cbb-upsets"
version = "0.1.0"
description = "NCAA men's basketball data ingest and storage CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "orjson>=3.10",
    "python-dotenv>=1.0.0",
    "requests>=2.32",
    "SQLAlchemy>=2.0",
    "psycopg2-binary>=2.9",
    "typer>=0.9,<1.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11",
    "pytest>=8.0",
    "ruff>=0.11",
]

[project.scripts]
cbb = "cbb.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
files = ["src"]
warn_unused_ignores = true
